refactor(chat): extract backend message mapper
