+++ CHANGELOG.md
@@ -44,6 +44,7 @@ The format is inspired by [Keep a Changelog](https://keepachangelog.com/en/1.1.0
 - Extracted timeline row rendering into `src/components/chat-view/message-timeline-view.ts` (message-loop orchestration + assistant/system/changelog/compaction row templates), reducing `chat-view` template density and consolidating timeline-specific rendering concerns.
 - Extracted slash built-in command orchestration into `src/components/chat-view/slash-builtin-command.ts`, including shared session-info block formatting, so `chat-view` now delegates built-in slash execution through a callback-driven runtime module.
 - Extracted git repository/branch picker rendering into `src/components/chat-view/git-repo-control-view.ts`, keeping `chat-view` focused on state orchestration while the branch menu template logic lives in a dedicated view module.
+- Extracted backend message normalization/mapping into `src/components/chat-view/backend-message-mapper.ts` and converted `chat-view` message mapping to a thin delegating adapter.
 - Compaction status rendering was reduced to a minimal workflow-style row with collapsed-by-default details instead of a heavy status card.
 - Auto-rename extension recommendation and desktop config bridge now target `@byteowlz/pi-auto-rename`, with dynamic command-to-package resolution for config-intent slash commands (including `/auto-rename config`) instead of hardcoded package-name routing.
 - Recommended notifications extension now defaults to `pi-smart-voice-notify`, and Packages auto-migrates legacy `pi-desktop-notify` installs by installing the new package and removing the old one.
